create reusable section component
